// product/src/modules/grpc/proto/v2.pb.ts
export interface GetUserResponse {
  id: string;
  name: string;
  phoneNumber: string;
  bankDetail: Bank | undefined;
  address: Address | undefined;
  email: string;
  isKeySeller: boolean;
  isMerchant: boolean;
  activeListings: number;
  soldListings: number;
  avatar: string;
  activatedDate: string;
  bio: string;
  isCompliant: boolean;
}

export interface GetDmUserResponse {
  id: string;
  firstName: string;
  lastName: string;
  username: string;
  status: string;
  email: string;
  phoneNumber: string;
}

export interface SellerOrderDetail {
  payoutAmount: number;
  sellPrice: number;
}

export interface BuyerOrderDetail {
  grandTotal: number;
}

export interface Bank {
  accountHolderName: string;
  accountId: string;
  bankBIC: string;
  bankName: string;
  isNonSaudiBank: boolean;
}

export interface Address {
  street: string;
  district: string;
  city: string;
  postalCode: string;
  latitude: string;
  longitude: string;
}